change batimage size fixed tongue position and added our menu for the variations to be played
Michael has helped me with fixing the size of the bat-image so that it is the size of the frog. Michael has also helped with the regular frogs tongue so it is not all droopy. ya Michael has also given me a shrimple solution to making my menu for the variations. Michael should be promoted from TA to dean of Concordia Michael should also get payed more than the other TA's, This Commit was "not" Endorse by Michael